refactor: tidy and depreciate
